StreamTime: Improve our stream timing logic.
Most of the time streams use an external reference clock which is
encoded in the bytestream of streaming data, take mpeg-ts as an example.
With these type of clocking, stream clock will be sync'ed to an external
clock at some intervals. For example:

......t1........t2........t3.........t4
.........x....y........................

The problem with using that external clock is that if elements queries
clock at points x and y they will see time as t1 in both cases which is
not correct at best and can be damaging for syncing at worst. So we have
a timer which is restarted at every reference points t(i) so at
reference points elements see the same time but increasing time values
at x and y.

We also add an free running timer which is re-started with every start
call and can be queried using getFreeRunningTime.