TextOverlay: Add support for a new method.
Our previous version was fast enough but it was painting a background
color onto the video. With this new method, we copy only painted pixels
in a character to the video so no background is drawn onto the image.
This method is a little bit slower when compared to previous one.