dmai: DmaiBuffer: New buffer class for DMAI-based operations.
TI's dmai library have its own buffer class and uses this class
all around the library for capturing, decoding, encoding etc. So
we need to hold references to dmai buffers in our own buffers. This
is becoming a very common operation so its best to create a new class
for this.

With this class we can use both implicit sharing and dmai buffers with
the same class.