RtspOutput: Use newly implemented ffmpeg based RTP muxer.
GStreamer based RTP streaming has memory leak problems, and I couldn't
find the root cause of this. Problem may be in GStreamer libraries or
our implementation. After restless days of war with this bug(), I'm
giving up and moving to an FFmpeg based implementation.

At the moment, this class needs a series clean-up, both gst and ffmpeg
based implementations are leaving with lots of code duplication and
it depends on both gstreamer and ffmpeg at build time.