H264Encoder: Re-work DmaiEncoder and H264Encoder implementations.
Up to now, DmaiEncoder was an abstract class and all inherited classes
should provide startCodec and friends.

However, this is not strictly inline with Codec Engine's IVIDENC1
interface. All sort of codecs implementing IVIDENC1 can be controlled
with IVIDENC1 APIs. This means that and MPEG-4 and H.264 encoder can be
controlled by same class. If you want to use codec specific
functionality then you are free to use codec specific structures in
IVIDENC1 APIs.

So with this commit DmaiEncoder becomes a generic IVIDENC1 compliant
encoder which can be used on its own and H264Encoder becomes H.264
specific implementation.