DM365CameraInput: Provide access to secondry image of vpbe.
VPBE resizer can provide a scaled version of input image. With this
commit we export this interface to our application. This way it will
be possible to provide a lower resolution stream output.

Kernel driver works in a strange way such that secondary stream output
is appended to the primary stream output. So we must provide driver one
buffer which will contain both images. But in our application we will
need these 2 images in 2 different buffers. For this reason we also
re-worked buffer handling in this class.

Now DM365CameraInput API provides another API called nextBuffer(int ch)
which gives access to secondary image buffers.