BaseLmmElement: Re-work threading support.
With this commit we remove 'threaded' variable and related API from
BaseLmmElement. Instead we assume all elements work in a threaded
environment from now on.

This change is necessary because current implementation may cause severe
and hard-to-catch bugs in the code. When we forget to set an element to
threaded very nasty things can occur. Besides we gain nothing with
current implementation, locking and extra mutex in input and output
buffer manipulation is a very fast operation so no need to look for an
extra variable.

BaseLmmOutput element is mostly effected from this change, because it
was managing its own threading mechanism which is plain-old wrong. Now
all thread code is removed from output elements, it is responsibility of
upper layers to use threads when needed.