H264Encoder: Add support for multi-pass encoding.
DM365 supports multi-pass encoding, well not quite but something
like that. They call this simple/pseudo multi-pass encoding. In
this mode we have 2 streams and 2 encoders. 1 stream is in low
resolution and the other one is in full resolution. Encoding
information coming from low-resolution stream is used by second
encoder to increase the quality of encoding.