BaseRtspServer: Add support for joining multicast sessions.
With this commit we add support for joining multicast sessions which
is the correct way of handling multicasts. Before this commit a new
multicast session is created with each new client but now we use
existing sessions if exists.

While joining multicast sessions we respect session names, meaning stream1m
and stream2m are different multicast sessions.