BaseLmmElement: Improve threading support.
This commit is work towards better threading support in library. Up to
now all elements was maintaining their own buffer management which was
causing severe problems.

The main problem is that we don't know what is the state of an element
during a stop operation. For example we don't:

* if it is in an blocked state,
* if it is processing an buffer
* if it is blocked how we will be unblocking it

and many more threading related questions. Moreover every elements
processing function was not well defined which can be improved a lot
with virtual functions. So what we do with this commit is:

* Make all input/output buffers and related locks private.
* Provide well defined protected functions for elements to manage
  their buffers.
* Add a pure virtual processBuffer() function which must be implemented
  by all BaseLmmElement sub-classes.
* Do not use one input buffer queue(list) and one output buffer queue(list)
  per element but permit any number of input/output queues to be created
  by elements and re-work addBuffer()/nextBuffer() functions to take
  multiple channels into account.
* Do not make addBuffer()/nextBuffer() functions virtual, derived classes
  will not implement these functions any more but implement  processBuffer()
  and friends instead.
* Implement proper element stoppping using these new features.
* Update BasePlayer to reflect new changes to its derived classes.
* Add proper blocking waiting function for output channels, i.e. one can
  sleep till all the buffers(or some of them) are processed by downstream
  elements. This is esspecially useful for elements reading from disks.

From now on we will handle buffer addition and retrieval in BaseLmmElement
which will give us predictable threaded start()/stop() behaviour for all
elements.

Please note that this commit is milestone commit and all applications should
be tested with this infrastructure. Nasty bugs may have been introduced.

These changes are tested on following platforms and it seems like
functionality is intact:

 * dm365
 * x86
 * dm8168

Following elements are most likely run-time broken with this commit, they
will need more work:

 * VideoTestSource