RawBuffer: Add a special buffer type EOF.
Players often need to know when the playback is finished and
situation is a bit messy at this point. Think about following
situtation:

	* Demux element reads 1000 buffers and returns ENOENT,
	* Demux thread quits.
	* Now you cannot stop playback because not all the frames
	  are decoded.
	* When all the elements finished, how are you going to tell
	  if it is EOF or waiting demux? Threads won't exit and
	  if you try to rely on buffer count of elements you will
	  fail because buffer propogation from one element to other
	  is not instantaneous.

The most reliable option here is to have an EOF buffer, which will
propogate into every element and related thread, if someone sees
this buffer she will definetely know what to do. This commit is
the RawBuffer part of this functionality, BaseLmmElement will follow.