tools: settings: Add incremental settings support.
Up to now, we were saving all settings in the system to disk, so that
they can be loaded on the next start by applications. This has one major
drawback such that you end up changing settings that you haven't touched
any time since they are saved to disk regardless of their changed state.

With this commit we add support for saving changes only when they are changed
so that only changed settings are loaded from disk on the next start. Old
behaviour is kept also.