Rtsp: BaseRtspServer: Add timeout support to base implementation.
We should close RTSP session after 60 seconds of inactivity, so
we add timeout support to RTSP with this commit. We also add support
for GET_PARAMETER command so that clients can ping server.

We also fix closeSession() function so that sessions closed by
RTCP or RTSP timeout are handled gracefully by derived classes.