RawBuffer: Use a class for buffer parameters instead of QHash.
By looking at profiling results, we see that addBufferParameter(),
bufferParameter() and related parameter functions take considerable
amount of time. So instead of using QHash we think that using a
dedicated class for buffer parameters can improve this performance hit.
Results show that this results in minimum 15% drop of CPU load on a
400 MHZ DM365(ARM9).

All classes effected from this change is modified as well. Moreover,
aboutDeleteBuffer() member of BaseLmmElement is effected from this change
so while we are at there, we renamed this function to aboutToDeleteBuffer
for correctness.