lmm: Big BaseLmmElement clean-up.
This clean-up is neccessary for upcoming better pipeline architecture
support for the lmm library. We are in process of integrating
pipelining support right into the heart of our library, BaseLmmElement.

BasePipeElement class is deprecated from now on, it is only
left around for existing builds. We will remove it once vksystem
is migrated to new APIs.