build: Prepare default build configuration.
Building all of our libraries and applications can be quite overwhelming for
mortal human beings, so we try to simplify things a little bit. Now it is
as easy as to compile lmm like this:

$git clone ...
$cd lmm/
$cp lmm/build_config.pri.sample lmm/build_config.pri
$cp dm365_ipstr/build_config.pri.sample dm365_ipstr/build_config.pri
$cp tests/build_config.pri.sample test/build_config.pri

... change to your desired build directory

$qmake /bla/bla/lmm/lmm-demo.pro
$make -j8 all
$make -j8 install

This will build lmm in your desired directory and install it into a directory
above the build directory. (../usr/local)

It is still possible to customize your build options using the build_config.pri.