dm365_ipstr: Add audio and mjpeg support.
With this commit we add audio and mjpeg streaming support. To be
compatiple with older RTSP scheming we introduce a new RTSP config
option and enable audio streaming in case new RTSP config scheme is
selected. This commit needs schema_version '7'.

All mjpeg and audio streaming can be disabled depending on config options.