rtp: RtpReceiver: Add a new feature we call bitrate summarization.
This feature is helpful for creating fancy bitrate graphs. Emitting a
signal every time we receive a buffer would make it impossible for
upper layers to track bitrate calcualtion because signals would stall
in the event loop. Instead we collect many packets into one and emit
signal with a saner period.