dm365_ipstr: Fix commit b8682bd0.
In commit b8682bd0164d1b638681b5ae71d564e1e22932bf we implemented pinger
in timeout() function which was wrong as hell. We should kick pinger
in pipeline output so that when we have a lock-up, we detect it. Current change
only checks lock-up for 0th  pipeline.

This change is only build tested.