BufferQueue: Fix a bug related to buffer order swapping.
Due to our pipeline implementation, processBlocking() function can be
called in multiple threads and our current buffer processing for this
element is not thread-safe. To be precised, buffer outputting logic
is not completely thread-safe so buffer orders in the output queue may be
swapped. This commit fixes this problem by only permitting one thread
to call processBuffer() function.